Make the app more restful
