+++ src/routes/__init__.py
@@ -1,0 +1,6 @@
+from routes.images import images_bp
+from routes.colours import colours_bp
+from routes.textures import textures_bp
+from routes.wang_tiles import wang_tiles_bp
+
+__all__ = ['images_bp', 'colours_bp', 'textures_bp', 'wang_tiles_bp']
